Add -json support for dry-run mode and -quiet flag to suppress logs
This commit introduces a `-json` flag that, when used together with
`-dry_run`, outputs the dry-run results in structured JSON format to
stdout.

Additionally, it adds a `-quiet` option that disables all log messages
sent to stderr, which is useful when a clean output is desired,
especially in combination with `-json`.

This facilitates integration with CI tools and scripts by providing a
structured output.

Example output:

{
  "dry_run_builds": [
    {
      "package": "foo",
      "version": "x.x-x"
      "sbuild_command": "sbuild..."
    },
    {
      "package": "bar",
      "version": "x.x-x"
      "sbuild_command": "sbuild..."
    }
  ]
}